add pre built package uploader options

+++ .github/scripts/python_package_upload/package_upload_run.sh
@@ -3,11 +3,20 @@
 set -ex
 
 CONTAINER_CLI="${CONTAINER_CLI:-docker}"
+PACKAGE_UPLOADER_IMAGE="${PACKAGE_UPLOADER_IMAGE:-}"
 
 mkdir -p /tmp/packages
-$CONTAINER_CLI rm package_upload_run || true
-$CONTAINER_CLI build -t package_upload .
-$CONTAINER_CLI run --name package_upload_run -v /tmp/packages:/app/packages package_upload
+
+if [ -z "$PACKAGE_UPLOADER_IMAGE" ]; then
+  echo "No prebuilt package uploader image specified, buildling..."
+  $CONTAINER_CLI rm package_upload_run || true
+  $CONTAINER_CLI build -t package_upload .
+  $CONTAINER_CLI run --name package_upload_run -v /tmp/packages:/app/packages package_upload
+else
+  echo "Using prebuilt package uploader image ${PACKAGE_UPLOADER_IMAGE}"
+  $CONTAINER_CLI run --name $PACKAGE_UPLOADER_IMAGE -v /tmp/packages:/app/packages package_upload
+fi
+
 # Print the pods in the namespace
 kubectl -n test-pypiserver get pods
